split properties panel out
